Rename users to snmp_users for compatibility with node-red 4.1.0
'users' became a restricted keyword per https://github.com/node-red/node-red/commit/26e58a53b720f741ca950e577dfbaf3cbe0d39a5
which causes an error when attempting to edit/save settings in node-red
4.1.0 and higher.

As a workaround, switch users to snmp_users, and switch communities to
snmp_communities for consistency.

A snippet has been added that runs at startup to migrate existing
nodes to the new config keys, along with a new 'schema' key to keep
track of if the migration is necessary. The updated config schema will
be 2, as the first was implicitly 1.

Bump version to 2.4.1 for release.